moved passport.mjs into middleware
